try fix timing problem

+++ Code2/Game.hpp
@@ -38,8 +38,8 @@ public:
 	Game(game_params);
 	~Game() = default; // TODO ??? SH
 	void run(); // Runs the game
-	const vector<float> gen_hist() const; // Returns the generation timing histogram  
-	const vector<float> tile_hist() const; // Returns the tile timing histogram
+	const vector<double> gen_hist() const; // Returns the generation timing histogram
+	const vector<double> tile_hist() const; // Returns the tile timing histogram
 	uint thread_num() const; //Returns the effective number of running threads = min(thread_num, field_height)
 
 
@@ -55,13 +55,14 @@ protected: // All members here are protected, instead of private for testing pur
     uint m_gen_num; 			 // The number of generations to run
 	uint m_thread_num; 			 // Effective number of threads = min(thread_num, field_height)
 	uint m_thread_temp;    // not Effective number of threads
-	vector<float> m_tile_hist; 	 // Shared Timing history for tiles: First m_gen_num cells are the calculation durations for tiles in generation 1 and so on.
+	vector<double> m_tile_hist; 	 // Shared Timing history for tiles: First m_gen_num cells are the calculation durations for tiles in generation 1 and so on.
 							   	 // Note: In your implementation, all m_thread_num threads must write to this structure. 
-	vector<float> m_gen_hist;  	 // Timing history for generations: x=m_gen_hist[t] iff generation t was calculated in x microseconds
+	vector<double> m_gen_hist;  	 // Timing history for generations: x=m_gen_hist[t] iff generation t was calculated in x microseconds
 	vector<Thread*> m_threadpool; // A storage container for your threads. This acts as the threadpool.
 
 	bool interactive_on; // Controls interactive mode - that means, prints the board as an animation instead of a simple dump to STDOUT 
 	bool print_on; // Allows the printing of the board. Turn this off when you are checking performance (Dry 3, last question)
+
 	vector<vector<int>> m_board;
 	vector<vector<int>> m_tmp_board;
 	vector<vector<int>> m_next_board;
