+++ Code2/main.cpp
@@ -2,7 +2,7 @@
 
 static inline game_params parse_input_args(int argc, char **argv);
 static inline void usage(const char* mes);
-static void calc_and_append_statistics(uint n_threads, const vector<float>& gen_hist, const vector<float>& tile_hist);
+static void calc_and_append_statistics(uint n_threads, const vector<double>& gen_hist, const vector<double>& tile_hist);
 
 /*--------------------------------------------------------------------------------
 										Main
@@ -46,7 +46,7 @@ static inline void usage(const char* mes) {
 }
 
 
-static void calc_and_append_statistics(uint n_threads, const vector<float>& gen_hist, const vector<float>& tile_hist) {
+static void calc_and_append_statistics(uint n_threads, const vector<double>& gen_hist, const vector<double>& tile_hist) {
 
 	float total_time = (float)accumulate(gen_hist.begin(), gen_hist.end(), 0.0);
 	float avg_gen_time = total_time / gen_hist.size();
